Make Config an instanced class and pass it through
Config wasn't being passed in correctly based on CLI options --
specifically, running `better-vsts-npm-auth -c ~/.myauthrc
config set foo bar` updates `~/.vstsnpmauthrc` rather than
`~/.myauthrc`. Additionally, it was hard to interact with
the API programmatically, due to the global state.

This change instantiates Config as part of the CLI initialization
flow, and modifies the function signatures of various functions
to take a Config in as a parameter.